add ColorCycle JSON encoder
